feat: add model_price integration tests and fix quota/RMB sync

+++ endpoints/openapi_v1/api_key/get_quota_plan.go
@@ -38,15 +38,15 @@ var GetQuotaPlanRoute = &xreq.Endpoint{
 type GetQuotaPlanResponse struct {
 	Unlimited             *bool    `json:"unlimited"`
 	PassWhenNoEnoughQuota *bool    `json:"pass_when_no_enough_quota"`
-	Quota                 *int64   `json:"quota"`
+	Quota                 *float64 `json:"quota"`
 	Unit                  *string  `json:"unit"`
 	ResetPeriod           *string  `json:"reset_period"`
 	Balance               *Balance `json:"balance"`
 }
 
 type Balance struct {
-	Used      int64 `json:"used"`
-	Remaining int64 `json:"remaining"`
+	Used      float64 `json:"used"`
+	Remaining float64 `json:"remaining"`
 }
 
 func GetQuotaPlanAction(req *http.Request) (interface{}, error) {
@@ -82,11 +82,11 @@ func GetQuotaPlanProcess(ctx context.Context, id string, product *ibasic.Product
 			return nil, err
 		}
 		if balanceData != nil {
-			used := int64(0)
+			used := float64(0)
 			if balanceData.Used != nil {
 				used = *balanceData.Used
 			}
-			remaining := int64(0)
+			remaining := float64(0)
 			if balanceData.Remaining != nil {
 				remaining = *balanceData.Remaining
 			}
@@ -95,7 +95,7 @@ func GetQuotaPlanProcess(ctx context.Context, id string, product *ibasic.Product
 				Remaining: remaining,
 			}
 		} else {
-			remaining := int64(0)
+			remaining := float64(0)
 			if apiKey.QuotaPlan.Quota != nil {
 				remaining = *apiKey.QuotaPlan.Quota
 			}
